+++ src/main.cpp
@@ -12,10 +12,8 @@ static_assert(INTPTR_MAX == INT64_MAX, "64-bit technology required: parameter-pa
 int main() {
     ///DEBUG_LOG_TO_FILE();
     srand(static_cast<int>(time(NULL)));
-#ifndef NDEBUG
-    CEngineTest::test_everything();
-#endif
     CUciProtocol uci_protocol;
+    // TODO: protocol-startup and engine-initialization before main message_loop
     uci_protocol.message_loop();
      return EXIT_SUCCESS;
 }