+++ src/technical_functions/docu_debug_log/working_simplified_demo/build.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/bash
-
-g++ -c DebugLog.cpp
-g++ -D DEBUG_LOG_ENABLE Program.cpp DebugLog.o